Moved common utils to shared library. Fixes #8
