+++ .gitignore
@@ -1,1 +1,3 @@
 __pycache__
+.vscode
+settings.json
